+++ src/main/java/com/umc/mada/custom/service/CustomService.java
@@ -38,7 +38,7 @@ public class CustomService {
         List<CustomItem> colorItems = customItems.stream().filter(item -> ItemType.I1.equals(item.getItemType())).collect(Collectors.toList());
         if(customItems.isEmpty()|| colorItems.isEmpty()){ //사용자 캐릭터가 디폴트 값이라면 디폴트 캐릭터 데이터를 넘겨준다.
             //return new UserCharacterResponse(CharacterItemResponse.of(customRepository.findCustomItemById(1L).get()));
-            customItems.add(customRepository.findCustomItemById(1L).get()); //TODO: isPresent() 체크하기
+            customItems.add(customRepository.findCustomItemById(1).get()); //TODO: isPresent() 체크하기
         }
         return UserCharacterResponse.of(customItems);
     }
@@ -69,7 +69,7 @@ public class CustomService {
         return customItemsResponse;
     }
 
-    public void changeUserItem(User user, Long item_id){
+    public void changeUserItem(User user, int item_id){
         Optional<CustomItem> customItem = customRepository.findCustomItemById(item_id);
         Optional<HaveItem> newhaveItemOptional = haveItemRepository.findByCustomItemAndUser(customItem.get(), user); //TODO: 예외처리 추가하기
         HaveItem newhaveItem = newhaveItemOptional.get();
@@ -111,7 +111,7 @@ public class CustomService {
         }
     }
 
-    public void buyItem(User user, Long item_id){
+    public void buyItem(User user, int item_id){
         Optional<CustomItem> customItemOptional = customRepository.findCustomItemById(item_id);
         CustomItem customItem = customItemOptional.get();
         if(haveItemRepository.findByCustomItemAndUser(customItem, user).isPresent()){
